float

+++ usatdata_erstellen_mistral.py
@@ -41,8 +41,8 @@ for m in month:
         cogfile='/mnt/lustre02/work/um0203/u301025/Masterarbeit/Eureka/Daten/Dship/seapath/2020'+m+str(d)+'cog_1Hz.dat'
         cog=pd.read_csv(cogfile, delimiter='\t', decimal='.',skiprows=[1,2] ,engine='python')
         print(seapathdata['SYS.CALC.SPEED_kmh'].dtype, cog['SYS.STR.Course'].dtype)
-        cog['X']=(seapathdata['SYS.CALC.SPEED_kmh'].astype(float)/3.6)*np.sin((float(cog['SYS.STR.Course']))/180. * math.pi)
-        cog['Y']=(seapathdata['SYS.CALC.SPEED_kmh'].astype(float)/3.6)*np.cos((float(cog['SYS.STR.Course']))/180. * math.pi)
+        cog['X']=(seapathdata['SYS.CALC.SPEED_kmh'].astype(float)/3.6)*np.sin(cog['SYS.STR.Course'].astype(float)/180. * math.pi)
+        cog['Y']=(seapathdata['SYS.CALC.SPEED_kmh'].astype(float)/3.6)*np.cos(cog['SYS.STR.Course'].astype(float)/180. * math.pi)
         # cog['date']=cog['date time'][:].astype('datetime64[ns]')
         cog['date']=pd.to_datetime(cog['date time'])
         cog.set_index(['date'],drop=True,append=False,inplace=True)
